【fix bug】1、rebuild时，当一个页面写满时加入到delete list会产生问题（认为是一个空页） 2、load/reorg等分配页面内容时当最小空间比minrecordsize小时，需要把这空间给上个record
git-svn-id: http://192.168.20.11/sequoiadb/trunk@10928 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dms/dmsReorgUnit.cpp
@@ -420,7 +420,7 @@ namespace engine
       recordPtr = ((ossValuePtr)currentExtent) + recordOffset ;
       if ( currentExtent->_freeSpace - (INT32)dmsrecordSize <
            DMS_MIN_RECORD_SZ &&
-           currentExtent->_freeSpace <= DMS_RECORD_MAX_SZ )
+           currentExtent->_freeSpace <= (INT32)DMS_RECORD_MAX_SZ )
       {
          dmsrecordSize = (UINT32)currentExtent->_freeSpace ;
       }
